requirements.txt
# Core dependencies
numpy>=1.26.0,<2.1.0   
opencv-python>=4.8.0       
Pillow>=12.0.0

# FastAPI stack
fastapi>=0.95
uvicorn>=0.22
starlette>=0.27
pydantic>=2.12
python-multipart>=0.0.20
anyio>=4.12.0
click>=8.3.1
h11>=0.16.0
idna>=3.11

# DICOM & ML
pydicom>=3.0.1
tensorflow==2.18.0


# Optional typing
annotated-doc>=0.0.4
annotated-types>=0.7.0
typing-inspection>=0.4.2
typing_extensions>=4.15.0
six>=1.17.0
